refactor: simplify engine manager grafloki skip logic
Replace method duplication approach with a struct field on EngineManager.
Instead of threading a shouldStartConfiguredGrafloki boolean through
3 private method copies + 1 new public method, use a skipConfiguredGrafloki
field with a setter. This removes ~100 lines of duplicated methods.

Also removes the logs_engine_restart helper package — loki/start and
grafloki/start now call the engine manager directly.
