+++ cli/cli/helpers/engine_manager/engine_manager.go
@@ -47,6 +47,7 @@ type EngineManager struct {
 	onBastionHost                             bool
 	enclaveEnvVars                            string
 	allowedCORSOrigins                        *[]string
+	skipConfiguredGrafloki                    bool
 	// Make engine IP, port, and protocol configurable in the future
 }
 
@@ -103,10 +104,11 @@ func NewEngineManager(ctx context.Context) (*EngineManager, error) {
 		kurtosisBackend:   kurtosisBackend,
 		shouldSendMetrics: kurtosisConfig.GetShouldSendMetrics(),
 		engineServerKurtosisBackendConfigSupplier: engineBackendConfigSupplier,
-		clusterConfig:      clusterConfig,
-		onBastionHost:      onBastionHost,
-		enclaveEnvVars:     enclaveEnvVars,
-		allowedCORSOrigins: nil,
+		clusterConfig:          clusterConfig,
+		onBastionHost:          onBastionHost,
+		enclaveEnvVars:         enclaveEnvVars,
+		allowedCORSOrigins:     nil,
+		skipConfiguredGrafloki: false,
 	}, nil
 }
 
@@ -116,6 +118,10 @@ func NewEngineManager(ctx context.Context) (*EngineManager, error) {
 //	commands only access the Kurtosis APIs, we can remove this.
 func (manager *EngineManager) GetKurtosisBackend() backend_interface.KurtosisBackend {
 	return manager.kurtosisBackend
+}
+
+func (manager *EngineManager) SetSkipConfiguredGrafloki(skip bool) {
+	manager.skipConfiguredGrafloki = skip
 }
 
 // GetEngineStatus Returns:
@@ -191,33 +197,6 @@ func (manager *EngineManager) StartEngineIdempotentlyWithDefaultVersion(
 	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
 	func() error, error,
 ) {
-	return manager.startEngineIdempotentlyWithDefaultVersion(
-		ctx,
-		logLevel,
-		poolSize,
-		githubAuthTokenOverride,
-		restartAPIContainers,
-		domain,
-		logRetentionPeriodStr,
-		additionalSinks,
-		true,
-	)
-}
-
-func (manager *EngineManager) startEngineIdempotentlyWithDefaultVersion(
-	ctx context.Context,
-	logLevel logrus.Level,
-	poolSize uint8,
-	githubAuthTokenOverride string,
-	restartAPIContainers bool,
-	domain string,
-	logRetentionPeriodStr string,
-	additionalSinks logs_aggregator.Sinks,
-	shouldStartConfiguredGrafloki bool,
-) (
-	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
-	func() error, error,
-) {
 	status, maybeHostMachinePortBinding, engineVersion, err := manager.GetEngineStatus(ctx)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred retrieving the Kurtosis engine status, which is necessary for creating a connection to the engine")
@@ -227,7 +206,7 @@ func (manager *EngineManager) startEngineIdempotentlyWithDefaultVersion(
 
 	var lokiSink logs_aggregator.Sinks
 	var grafanaUrl string
-	if shouldStartConfiguredGrafloki && manager.clusterConfig.GetGraflokiConfig().ShouldStartBeforeEngine {
+	if !manager.skipConfiguredGrafloki && manager.clusterConfig.GetGraflokiConfig().ShouldStartBeforeEngine {
 		lokiSink, grafanaUrl, err = grafloki.StartGrafloki(ctx, clusterType, manager.clusterConfig.GetGraflokiConfig())
 		if err != nil {
 			return nil, nil, stacktrace.Propagate(err, "An error occurred starting Grafana and Loki before engine.")
@@ -285,36 +264,6 @@ func (manager *EngineManager) StartEngineIdempotentlyWithCustomVersion(ctx conte
 	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
 	func() error, error,
 ) {
-	return manager.startEngineIdempotentlyWithCustomVersion(
-		ctx,
-		engineImageVersionTag,
-		logLevel,
-		poolSize,
-		shouldStartInDebugMode,
-		githubAuthTokenOverride,
-		restartAPIContainers,
-		domain,
-		logRetentionPeriodStr,
-		additionalSinks,
-		true,
-	)
-}
-
-func (manager *EngineManager) startEngineIdempotentlyWithCustomVersion(ctx context.Context,
-	engineImageVersionTag string,
-	logLevel logrus.Level,
-	poolSize uint8,
-	shouldStartInDebugMode bool,
-	githubAuthTokenOverride string,
-	restartAPIContainers bool,
-	domain string,
-	logRetentionPeriodStr string,
-	additionalSinks logs_aggregator.Sinks,
-	shouldStartConfiguredGrafloki bool,
-) (
-	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
-	func() error, error,
-) {
 	status, maybeHostMachinePortBinding, engineVersion, err := manager.GetEngineStatus(ctx)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred retrieving the Kurtosis engine status, which is necessary for creating a connection to the engine")
@@ -324,7 +273,7 @@ func (manager *EngineManager) startEngineIdempotentlyWithCustomVersion(ctx conte
 
 	var lokiSink logs_aggregator.Sinks
 	var grafanaUrl string
-	if shouldStartConfiguredGrafloki && manager.clusterConfig.GetGraflokiConfig().ShouldStartBeforeEngine {
+	if !manager.skipConfiguredGrafloki && manager.clusterConfig.GetGraflokiConfig().ShouldStartBeforeEngine {
 		lokiSink, grafanaUrl, err = grafloki.StartGrafloki(ctx, clusterType, manager.clusterConfig.GetGraflokiConfig())
 		if err != nil {
 			return nil, nil, stacktrace.Propagate(err, "An error occurred starting Grafana and Loki before engine.")
@@ -464,69 +413,6 @@ func (manager *EngineManager) RestartEngineIdempotently(ctx context.Context,
 	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
 	func() error, error,
 ) {
-	return manager.restartEngineIdempotently(
-		ctx,
-		logLevel,
-		optionalVersionToUse,
-		restartEngineOnSameVersionIfAnyRunning,
-		poolSize,
-		shouldStartInDebugMode,
-		githubAuthTokenOverride,
-		shouldRestartAPIContainers,
-		domain,
-		logRetentionPeriodStr,
-		additionalSinks,
-		true,
-	)
-}
-
-func (manager *EngineManager) RestartEngineIdempotentlyWithoutConfiguredGrafloki(ctx context.Context,
-	logLevel logrus.Level,
-	optionalVersionToUse string,
-	restartEngineOnSameVersionIfAnyRunning bool,
-	poolSize uint8,
-	shouldStartInDebugMode bool,
-	githubAuthTokenOverride string,
-	shouldRestartAPIContainers bool,
-	domain string,
-	logRetentionPeriodStr string,
-	additionalSinks logs_aggregator.Sinks,
-) (
-	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
-	func() error, error,
-) {
-	return manager.restartEngineIdempotently(
-		ctx,
-		logLevel,
-		optionalVersionToUse,
-		restartEngineOnSameVersionIfAnyRunning,
-		poolSize,
-		shouldStartInDebugMode,
-		githubAuthTokenOverride,
-		shouldRestartAPIContainers,
-		domain,
-		logRetentionPeriodStr,
-		additionalSinks,
-		false,
-	)
-}
-
-func (manager *EngineManager) restartEngineIdempotently(ctx context.Context,
-	logLevel logrus.Level,
-	optionalVersionToUse string,
-	restartEngineOnSameVersionIfAnyRunning bool,
-	poolSize uint8,
-	shouldStartInDebugMode bool,
-	githubAuthTokenOverride string,
-	shouldRestartAPIContainers bool,
-	domain string,
-	logRetentionPeriodStr string,
-	additionalSinks logs_aggregator.Sinks,
-	shouldStartConfiguredGrafloki bool,
-) (
-	kurtosis_engine_rpc_api_bindings.EngineServiceClient,
-	func() error, error,
-) {
 	var versionOfNewEngine string
 	// We try to do our best to restart an engine on the same version the current on is on
 	_, _, currentEngineVersion, err := manager.GetEngineStatus(ctx)
@@ -551,9 +437,9 @@ func (manager *EngineManager) restartEngineIdempotently(ctx context.Context,
 	var engineClientCloseFunc func() error
 	var restartEngineErr error
 	if versionOfNewEngine != defaultEngineVersion {
-		_, engineClientCloseFunc, restartEngineErr = manager.startEngineIdempotentlyWithCustomVersion(ctx, versionOfNewEngine, logLevel, poolSize, shouldStartInDebugMode, githubAuthTokenOverride, shouldRestartAPIContainers, domain, logRetentionPeriodStr, additionalSinks, shouldStartConfiguredGrafloki)
+		_, engineClientCloseFunc, restartEngineErr = manager.StartEngineIdempotentlyWithCustomVersion(ctx, versionOfNewEngine, logLevel, poolSize, shouldStartInDebugMode, githubAuthTokenOverride, shouldRestartAPIContainers, domain, logRetentionPeriodStr, additionalSinks)
 	} else {
-		_, engineClientCloseFunc, restartEngineErr = manager.startEngineIdempotentlyWithDefaultVersion(ctx, logLevel, poolSize, githubAuthTokenOverride, shouldRestartAPIContainers, domain, logRetentionPeriodStr, additionalSinks, shouldStartConfiguredGrafloki)
+		_, engineClientCloseFunc, restartEngineErr = manager.StartEngineIdempotentlyWithDefaultVersion(ctx, logLevel, poolSize, githubAuthTokenOverride, shouldRestartAPIContainers, domain, logRetentionPeriodStr, additionalSinks)
 	}
 	if restartEngineErr != nil {
 		return nil, nil, stacktrace.Propagate(restartEngineErr, "An error occurred starting a new engine")
